start extracting frontend task client, add more code comments
